Move declarations to cpp file (apparent_safe_velocity_limiter_node)
