--- planning/apparent_safe_velocity_limiter/src/apparent_safe_velocity_limiter.cpp
@@ -1,22 +1,0 @@
-// Copyright 2022 Tier IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#include "apparent_safe_velocity_limiter/apparent_safe_velocity_limiter_node.hpp"
-
-namespace apparent_safe_velocity_limiter
-{
-}  // namespace apparent_safe_velocity_limiter
-
-#include "rclcpp_components/register_node_macro.hpp"
-RCLCPP_COMPONENTS_REGISTER_NODE(apparent_safe_velocity_limiter::ApparentSafeVelocityLimiterNode)
